requirements.txt
fastapi
uvicorn
flask
pytest
requests
boto3
python-dotenv==0.19.2
